+++ library/Episciences/Submit.php
@@ -996,13 +996,14 @@ class Episciences_Submit
 
         if ($isCoiEnabled) {
 
-            $cUidS = Episciences_Paper_ConflictsManager::fetchSelectedCol('by', ['answer' => 'no', 'paper_id' => $paper->getPaperid()]);
+            // conflict UIDs
+            $cUidS = Episciences_Paper_ConflictsManager::fetchSelectedCol('by', ['answer' => Episciences_Paper_Conflict::AVAILABLE_ANSWER['no'], 'paper_id' => $paper->getPaperid()]);
 
             foreach ($recipients as $recipient) {
 
-                $rUid = $recipient->getUid();
+                $rUid = $recipient->getUid(); // current uid
 
-                if (!in_array($rUid, $cUidS, false)) {
+                if (array_key_exists($rUid, $cUidS)) {
                     unset($recipients[$rUid]);
                 }
             }