+++ library/Episciences/Submit.php
@@ -207,7 +207,16 @@ class Episciences_Submit
             if (array_key_exists('canPickEditors', $settings) && $settings['canPickEditors'] > 0) {
                 // Récupération et tri des valeurs du select
                 $options = [];
-                $users = Episciences_UsersManager::getUsersWithRoles([Episciences_Acl::ROLE_CHIEF_EDITOR, Episciences_Acl::ROLE_EDITOR]);
+
+                $with = (
+                    array_key_exists(Episciences_Review::SETTING_DO_NOT_ALLOW_EDITOR_IN_CHIEF_SELECTION, $settings) &&
+                    $settings[Episciences_Review::SETTING_DO_NOT_ALLOW_EDITOR_IN_CHIEF_SELECTION]
+                )
+                    ? Episciences_Acl::ROLE_EDITOR :
+                    [Episciences_Acl::ROLE_CHIEF_EDITOR, Episciences_Acl::ROLE_EDITOR];
+
+
+                $users = Episciences_UsersManager::getUsersWithRoles($with);
 
                 /* @var  $user Episciences_User */
                 foreach ($users as $uid => $user) {
@@ -987,13 +996,14 @@ class Episciences_Submit
 
         if ($isCoiEnabled) {
 
-            $cUidS = Episciences_Paper_ConflictsManager::fetchSelectedCol('by', ['answer' => 'no', 'paper_id' => $paper->getPaperid()]);
+            // conflict UIDs
+            $cUidS = Episciences_Paper_ConflictsManager::fetchSelectedCol('by', ['answer' => Episciences_Paper_Conflict::AVAILABLE_ANSWER['no'], 'paper_id' => $paper->getPaperid()]);
 
             foreach ($recipients as $recipient) {
 
-                $rUid = $recipient->getUid();
+                $rUid = $recipient->getUid(); // current uid
 
-                if (!in_array($rUid, $cUidS, false)) {
+                if (array_key_exists($rUid, $cUidS)) {
                     unset($recipients[$rUid]);
                 }
             }