+++ application/languages/en/views.php
@@ -2248,4 +2248,17 @@ return [
 
     Episciences_Mail_TemplatesManager::TEMPLATE_DESCRIPTION_AND_RECIPIENT[Episciences_Mail_TemplatesManager::TYPE_PAPER_FORMATTED_BY_JOURNAL_WAITING_AUTHOR_VALIDATION][Episciences_Mail_TemplatesManager::DESCRIPTION] => "notification inviting the author to confirm the version to the journal's standards (formatted by the copy preparer)",
 
+    // Mot de passe papier
+    'Mot de passe papier' => 'Paper password',
+    'Afficher le mot de passe' => 'Show Password',
+    "Le mot de passe papier n'a pas été enregistré" => 'Paper password has not been saved',
+    'le champ est vide.' => 'the field is blank.',
+    'car il est identique à celui déjà enregistré.' => 'because it is identical to the one already registered.',
+    'Votre mot de passe papier a bien été enregistré.' => 'Paper password has been saved.',
+    "Si vous le souhaitez, vous pouvez partager le mot de passe papier permettant ainsi à la revue de mettre à jour ou de remplacer ce papier dans l'archive ouverte." => 'If you wish, you can share the paper password allowing the journal to update or replace this paper in the open archive.',
+    'L’auteur peut déléguer à la revue la mise à jour ou le remplacement de sa soumission publiée sur arXiv' => 'The author can delegate to the journal the updating or replacement of his submission published on arXiv',
+    'Copier' => 'Copy',
+    "Permettre aux auteurs de partager le mot de passe papier arXiv" => "Contributors can share arXiv paper password"
+
+
 ];