+++ application/modules/common/controllers/DefaultController.php
@@ -254,7 +254,7 @@ class DefaultController extends Zend_Controller_Action
 
         if ($isCoiEnabled) {
 
-            $cUidS = Episciences_Paper_ConflictsManager::fetchSelectedCol('by', ['answer' => 'no', 'paper_id' => $paper->getPaperid()]);
+            $cUidS = Episciences_Paper_ConflictsManager::fetchSelectedCol('by', ['answer' => Episciences_Paper_Conflict::AVAILABLE_ANSWER['no'], 'paper_id' => $paper->getPaperid()]);
 
             foreach ($recipients as $recipient) {
                 $rUid = $recipient->getUid();