+++ CHANGELOG.md
@@ -16,11 +16,15 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 -->
 
 ## Unreleased
+### Added
+- It is now possible to use Lemon LDAP authentication.
+- new option: do not allow the selection of an editor in chief when the author has the option to propose an editor at the time of submission.
 ### Changed
 - [#255](https://github.com/CCSDForge/episciences/issues/255)
 
-### Added
-- It is now possible to use Lemon LDAP authentication.
+### Fixed
+- [RT#169088]: Editors in chief and editorial secretaries are no longer notified of new submissions.
+
 
 ## 1.0.34.2 - 2022-10-18
 ### Fixed