+++ CHANGELOG.md
@@ -15,6 +15,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Security
 -->
 
+## Unreleased
+### Changed
+- Optimization: there is no need to check for conflicts at the time of submission
+
 ## 1.0.35 - 2022-10-19
 ### Added
 - [RT#169107]: Feature - new option: do not allow the selection of an editor-in-chief when the author has the option to