+++ CHANGELOG.md
@@ -16,6 +16,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 -->
 
 ## Unreleased
+### Added
+- authors can now enter their paper password provided by the open archive on the article page.
+
 ### Changed
 - Optimization: there is no need to check for conflicts at the time of submission